refactor: Don't Populate __init__.py
