+++ server/__main__.py
@@ -1,3 +1,7 @@
+import logging
+import sys
+from server.common.utils.utils import import_plugins
+
 # Work around bug https://github.com/pallets/werkzeug/issues/461
 if __package__ is None:
     import sys
@@ -11,5 +15,12 @@ if __package__ is None:
 
 # Main thing
 from .cli.cli import cli  # noqa F402
+
+try:
+    import_plugins("server.plugins")
+except Exception as e:
+    # Make sure to exit in this case, as the server may not be configured as expected.
+    logging.critical(f"Error in import_plugins: {str(e)}")
+    sys.exit(1)
 
 cli()
