+++ server/common/config/client_config.py
@@ -1,5 +1,4 @@
-from server import display_version as cellxgene_display_version
-
+from server.common.utils.version import display_version as cellxgene_display_version
 
 def get_client_config(app_config, data_adaptor, current_app):
     """
