+++ server/common/utils/version.py
@@ -1,0 +1,2 @@
+__version__ = "0.16.7"
+display_version = "cellxgene v" + __version__
